hotfix: eliminate extraneous layer deps
Right now cannon has a check for extraneous dependencies. unfortunately,
when assembling layers for a cannon chart, it is possible that
it can accumulate extraenous deps within the layers even if the original
cannonfile was clean.

this process is very slow: in worst case, the algo is now cubic speed, (
though this isnt taking into account ridiculous cases). If the user
has a modern comture, the layer calculation should still be reasonably
fast, but we will want to be looking for ways to optimize this algorithm
in the future.

Keeping in the checks to prevent extraneous dependencies should help,
but we will probably need a system to track "newly" extraneous dependencies
so that the search can be much faster, if needed