even better error handling, support for overriding hardhat config
one of the problems is that many plugins fail to work properly with cannon
because its not properly overriding the hardhat configuration.

not anymore!

the only requirement is to adjust the `defaultNetwork` property to be
`defaultNetwork: 'cannon'`. `cannon` is a well-known name recognized by the hardhat-cannon
project which will cause it to recognize the network as used for snapshotting/forks.

this fixes plugins like `hardhat-gas-reporter` which depend on usage of the hre for provider information

also I tried to implement error handling, but I think some aspects of it may still
be malfunctioning. Its fine if you use the provider returned by cannon, but
you may run into problems if you don't use it.

Finally, it may be a little awkward how the `CannonWrapperProvider` (or whatever its called)
uses a field `artifacts` to determine the decoding configuration for errors. It may be better
to have an actual setter function instead.