fix isolated skip context not carried over
during package upgrades, if a step has no changes in the check hash, the step will be skipped
even if dependencies had changed. This is desirable behavior. However, if the context of a upstream dependency
has changed, it is important for that change to be carried over to subsequent contexts.

We can accomplish this by simply overriding the previous context onto the new context. This is acceptable
since context can only extended by subsequent steps (aka, not upgraded).