add capability to grant additional deployers
the owner of a package may grant a list of additional deployers.
these deployers can publish to a package as if they are the owner.

this is useful if you want to have multiple parties deploying a package, and
want to lower the security level of publishing a package (since deployers
do not have permission to nominate a new owner)