add PreContext
in order to allow for settings to be interpolated with results from the
package (or timestamp or etc), a "pre" context is required. The pre
context just includes properties which are able to be resolved prior to
building a package, such as chainId, timestamp, and package.json contents.

may be helpful to update the documentation to explain exactly what is in
the context during settings/version resolution, but the fact that other
properties (such as the value of "other" settings) is not yet resolved.