default `targetPreset` to a different context
changes the default `targetPreset` from `main` (which causes problems
and overwrites package versions and etc) to `with-${packageName}`, where
`${packageName}` is the name of the currently being built package.

unfortunately, we currently have no way to know what package is
currently being built, so the `currentLabel` which is ususally passed
into actions has been augmented to support both the package name and
version as well. The builder will automatically handle any organization
of this.