edge case: handling of upstream artifacts and other misc data
in the case of a partial deployment, it was possible that misc data
could be lost when a step is not reexecuted, yet the data from upstream
misc is reloaded.

To fix this, ensure misc is separately maintained for provisioned
deployments, and in the spirit of having template to work off of,
`getArtifact` actually returns the misc data from upstream as required.

profit!