use multicall on copy package publish
recently more and more I have been finding that publish transactions
tend to be very numerous, which is both expensive and arduous. It would
be better if package publish calls were the minimum number possible.

Add a `publishMany` function to the CannonRegistry (which defaults to
just calling `publish` in a loop. This function is overridden by
OnChainRegistry, which will combine all the calls into a single
multicall transaction