support for recursive verify
now all provisioned packages will be eligible for publish upon call to
`cannon verify`.

This is achieved with a new function `forPackageTree` which basically
executes an action for a given `DeploymentInfo` object recursively, one
at a time.

We could also use this function to do all kinds of other things to do
with subpackages.