shared signals state for all derived runtimes
previously, if executing a build on a gitops repo or anything that uses
nested provisioning. the cancelling of the build will fail because the `cancelled`
boolean only gets applied to the top level runtime, but it needs to be shared between all runtimes

the way that this sharing is accomplished is by using shared reference of a "signals" object.

In the future if we add any additional signals then they can go in there as well.