fix infinite loop issue with layer compute
the problem is that the layer resolution system did not just depend on
having topological dependencies. As an additional requirement, steps
should be arranged based on the *layer* that they appear on to prevent
cases where a subdependency shared by two steps ends up being merged
together on a lower layer, creating a internal dependency cycle.

thankfully, it is fairly easy to mitigate this issue. if, while scanning
deep dependencies, we also return the `maxDepth` (which is the effective
depth of the dependency that was just scanned) and recursively compute
that, we can assign a layer depth to every action reasonably cheaply.
After this, it is possible to sort all nodes by depth (nodes with higher
depth are mathematically guarenteed to have all dependencies included
over nodes with a lower depth). This also gives us the guarentee we need
in order to properly build the merged layer tree, required for local
builds.