add support for cannon trace command
implements `cannon trace`, which allows for the generation of a **stack
trace** of *any* transaction on chain, or a specified transaction code
offchain.

the cannon trace command works similar to the `cannon decode` command.
Simply supply a package name and hex string.

can be used to decode any transaction already on chain by submitting the
transaction has as the hex string. The exact parameters under which the
original transaction executed will be accomodated in the simulation to
ensure that the result is exactly the same as its original circumstance
(individual fields, such as `from` or `block`, can be overridden if
desired)

alternatively, if the transaction isnt submitted on-chain or you want to
trace a *read-only* call, supply the exact call data of the transaction
you want to execute. Cannon will auto-detect the contract to send the
call to and submit the transaction automatically. To override the
destination contract (as well as other options such as `value` or
`from`), those values can be supplied as needed.