fix: incorrect handling of providerUrl
for some reason the eth-provider library does not properly handle
explicit `--provider-url` when they are specified, at least for
localhost forks which are created for a manual test. When this occurs,
it becomes impossible to specify a local fork as a target network for a
test. this error probably only is relevant when the user is running a
foundry instance (and goes away as soon as the foundry instance is shut
down).

To solve this problem, bypass eth-provider ocmpletely when an explicit
url is provided by the user.