fix(cli): much faster startup
often cannon cli would startup slow because the `npm view` command that
is called to resolve the latest cannon version has to call the internet
to get the latest version. Its not necessary to call out to the internet
every time the user uses cannon to check for the latest version, just
very periodically.

to solve this problem, the `npm view` command is called only once a week
at max. in other cases, the version is loaded from a file stored in
their cannon directory.

on my computer this brings the startup time of the cannon process from
1-2 seconds to start every time to near instant.