fix(builder): file links in recursive provision packages
it is possible that a package which has 2 levels deep of provision
artifacts can preserve a rogue file link after being uploaded to ipfs

to prevent this, change the functionality of the url remapping in the
forPackageTree function to support loading of chain artifacts.

this is kind of an expensive change, but we gotta do what we gotta do
right?

i have tested this independently with the sample-hardhat-cannon project,
and it worked as expected both for ipfs storage, and for file storage