feat: better handling of contract source codes omission
previously, we used envvars/cannon config to decide whether or not to include contract source codes in the cannon data.

instead, now the storage of cannon source codes is determined by a global property in the cannonfile. this should solve all cases of determining whether or not to deploy
contract sources safely, in roughly the same way that the npm package.json `private` key works