feat(website): Add getExplorerUrl and formatToken helpers.
This commit adds a new function `getExplorerUrl` to the `chains.ts` file in the `helpers` directory. This function takes a `chainId` and a `hash` as parameters and returns the corresponding explorer URL for the given chain and hash.

In addition, a new file `index.ts` is added to the `formatters` directory, which exports a `formatToken` function. This function takes an `amount` and an optional `options` object as parameters and returns a formatted token string.