ability to load
this is necessary to support some edge cases introduced by chains such as arbitrum, which do not load any actual code into their chain state so forking fails.

the idea is to load moc contracts which will be sufficient for most cases of forking.

the first precompile that has been added is the arbitrum gas oracle contract. the rest of the arbitrum precompiles should be added in the near future.