require `--impersonate` or `--impersonate-all` on dry run builds
its an extremely common error that unexpected signers are needed to
complete a deployment, so this will ensure that a user has to be
explicit about which signers should be required.

this is a breaking change, of course. Roll it in with all the others
coming out in the next release.

TODO: do we need to check if `--private-key` is provided also? if it is
then it could be treated as an implicit `--impersonate`