add field to get current deployer from inside cannon context
should return the signer that will be used if `getDefaultSigner()` is called. Note that this default signer can implicitly be a different signer depending on the txn, so there is no 100% way to reconcile this, but many times, its the same signer no matter the txn (unless explicitly defined otherwise).

if a static deployer is available, it will be supplied as `<%= deployer %>`.

if a static deployer is not available, it is set to undefined.

a bit of research/testing has to be done to determine the consistency of using this field, because, similar to `timestamp`, it can cause cannon package build to be undeterministic/unstable